+++ Velvet/entlist.h
@@ -9,7 +9,7 @@ public:
 	{
 		return getvfunc<centity *(__thiscall *)(void *, int)>(this, 3)(this, i);
 	}
-	inline centity *entfromhandle(unsigned long i)
+	inline centity *entfromhandle(DWORD i)
 	{
 		return getvfunc<centity *(__thiscall *)(void *, int)>(this, 6)(this, i);
 	}
